tries dirt fix

+++ src/actuators/motorscontrol.h
@@ -69,7 +69,7 @@ public:
     void updateWheelsPositions();
     double currentLimiter(double normalizedWheelSpeed, double power);
 
-    double getAngleError(double desiredAngle, double realAngle);
+    int getAngleError(double desiredAngle, double realAngle);
     double getNewAngle();
     double getNewPosition();
     double getNewRightWheelPosition();
